feat: Add classes for files and folders
redo structure

Issue #27
